Fixed Passpoint network "saved" order in wifi picker
Changed AccessPoint.isSaved() to check that mConfig is not null,
instead of checking if the networkId is valid. The networkId is set
to the networkId of the loaded config, but the configs received
for Passpoint networks do not have a valid networkId set. Thus,
Passpoint AccessPoints were not treated as saved during sorting,
despite showing "Saved" in the summary.

Test: atest WifiTrackerTest && atest AccessPointTest
Bug: 118705403
Change-Id: I254602aee8852a2acbf0c987a31051f2b9ae2a92