Fix janky on permission bar chart
In old design, the height of bar view is wrap_content.
So, it may have additional space after we normalized the
height for each bar view.
As a result, the button moves down when permission chart
gets filled.

For now, we fix the height which is max normalized height
for bar view. So, it won't have additional space while widget
is rendering the chart.

Test: See bar chart in privacy settings and permission dashboard.
Fixes: 123539793
Change-Id: Idb2619e21606248079715e055cb26ff366c71ad3