Fix partically truncated problem for bar chart
- We fix height of detail view, so we see partially
truncated when the text wraps to next line.
For now, we set the height of detail view to wrap_content.
And then we initilize title and detail view earlier
so that we won't get a jankey problem.

- We enlarge the max height of bar view so that we
won't get truncated problem when user sets display and
font size to largest.
We will still see truncated problem in language, display size,
font size combo. But we only optimize for majority use cases.

Fixes: 126266231
Fixes: 126720230
Fixes: 126726034
Test: See bar chart in Settings app and Permission controller.

Change-Id: I08965c715d868cee055b89c6d56930f3c2ae6714