Make Paint default to filtering Bitmaps
Bug: 77949917
Test: I5737be68dccdc5ebb38291bf3515ae25553dcfc2

In P, the switch to Skia HW rendering lost a feature of HWUI's old
renderer: using a Paint with a BitmapShader resulted in filtering the
Bitmap, regardless of whether FILTER_BITMAP_FLAG was set.

This somewhat restores that behavior, in that it results in filtering
BitmapShaders so long as the client doesn't explicitly turn it off. This
also means that it continues to be possible to get a pixellated drawing
if the client *does* explicitly turn it off.

Change-Id: I62c726a0ba281de239092f066a3df041ecbcee5b