Allow touchable regions for assist UI
System vendors may require their system UI to show UI for an assistant.

This CL adds methods to AssistManager to communicate if the system UI
is showing UI for the assistant, as well as a region of the screen that
accepts touches for the assistant UI.

Test: Implemented subclass of AssistManager with real implementations
of the new methods. Verified touches were intercepted/not intercepted
as setShowingAssistUi(false/true) was called.

Test: verified Assist UI is presented/disappears as appropriate
Change-Id: I7702f80b75a231dc5718128e49335d34beeaccfd