Reland "Make Paint default to filtering Bitmaps"
Bug: 77949917
Test: Iac4922cfa86c0c2b3b6c88b39d9b702ab046cd34

In P, the switch to Skia HW rendering lost a feature of HWUI's old
renderer: using a Paint with a BitmapShader resulted in filtering the
Bitmap, regardless of whether FILTER_BITMAP_FLAG was set.

This somewhat restores that behavior, in that it results in filtering
BitmapShaders so long as the client doesn't explicitly turn it off. This
also means that it continues to be possible to get a pixellated drawing
if the client *does* explicitly turn it off.

This reverts commit cf10d42469651082c06f90a4babd6b8e4eb71748. Original
CL was reverted due to a broken CTS test, which will be updated along
with this.

Change-Id: I61394b4ca6b93ba8405a0127b9c308444abfcbc2