Revert "MediaHTTPConnection: move connection states into an inner class"
This reverts commit ca56af8b157a9416d14a00f69a676fae8a187999.

Many of the fields that were moved are annotated @UnsupportedAppUsage,
so the CL would have had undesirable app compat impact. Further,
because investigation has revealed that lock contention *is* possible,
we need to always acquire the lock anyway so there is no longer a
benefit in keeping all of the mutable state in a single field that
can be atomically set to null.

Test: Treehugger
Change-Id: I202c5653cb086d99228491e161a159bad640105a