Ensure all fields of AutoBufferPointer are initialized
In the process of revising libnativehelper interface, the
initialization of the fArray field was dropped. This change fixes
this and adds regression tests.

In writing the tests, it became apparent that the AutoBufferPointer
destructor would crash if the underlying buffer was heap array backed
and the buffer position was non-zero. ReleasePrimitiveArrayCritical
was called with the element corresponding to the current position
rather than the first element.

Bug: b/130390512
Test: atest FrameworksCoreTests:android.graphics.BitmapTest

(cherry picked from commit f21e0c8d27bed10177f96b6c328f07c065005b9d)

Merged-In: I319512aaede7ba8af5d013c9281417695abf9099
Change-Id: I319512aaede7ba8af5d013c9281417695abf9099