BubbleData [7/n]: BubbleData impl and tests!!
This adds the full implementation of BubbleData along with
test cases developed against the designed behavior of Bubbles.

See: go/bubble-order for the full policy and rules behind this.

Since BubbleData.Listener#onOrderChanged is not yet connected,
the behavior is not yet applied to UI. This will happen in a
following change (along with new BubbleStackView tests).

Bug: 123542488
Test: atest BubbleControllerTest BubbleDataTest
Change-Id: I79d4452c196945735f081d6ae0fdc673de7ae102